requirements.txt
fastapi[all]
pydantic-settings==2.9.1
pyjwt==2.10.1
passlib==1.74
google-auth==2.39.0
sqlmodel==0.0.24
sqlalchemy==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
gTTS==2.5.4
faster-whisper==1.1.1
numpy==1.26.4
openai==1.75.0
pandas==2.2.3
pyparsing==3.1.4
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
torch==2.5.0
tqdm==4.66.5